*move MergePermute, MergePooling, MergeSpaceToDepth, MergeNormalize, MergeScale, MergeTiledScale2D and MergeUnpack4 to separate source files.
