+++ prj/vs2022/CvtCore.vcxproj
@@ -34,12 +34,19 @@
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeCurrentAndBias.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeInnerProductAndPower.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeInnerProductAndScale.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeNormalize.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergePermute.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergePooling.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergePowerAndScaleAndPower.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantized.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeQuantizedSqueezeExcitation.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeScale.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeShuffle.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSoftmax.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSpaceToDepth.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\MergeSqueezeExcitation.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeTiledScale2D.cpp" />
+    <ClCompile Include="..\..\src\Cvt\Optimizer\MergeUnpack.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\Optimizer.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\ReduceTensorIteratorIO.cpp" />
     <ClCompile Include="..\..\src\Cvt\Optimizer\TransposeInnerProduct.cpp" />
