+++ src/Cvt/Optimizer/Common.h
@@ -93,6 +93,26 @@ namespace Synet
 
     bool MergeShuffle4(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes);
 
+    bool MergePermute(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes);
+
+    bool MergePooling(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes);
+
+    bool MergeSpaceToDepth(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes);
+
+    bool MergeNormalize(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes);
+
+    bool MergeNormalizeV2(const LayerParams& src, size_t& index, bool isNhwc, LayerParams& dst, Changes& changes);
+
+    bool MergeNormalizeV4(const LayerParams& src, size_t& index, bool isNhwc, LayerParams& dst, Changes& changes);
+
+    bool MergeNormalizeV5(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes);
+
+    bool MergeScale(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes);
+
+    bool MergeTiledScale2D(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes);
+
+    bool MergeUnpack4(const LayerParams& src, size_t& index, bool isNhwc, LayerParams& dst, Changes& changes);
+
     bool MergeConvolutionAndScale(const LayerParams& src, size_t& index, const Bytes& bin, Bytes& buf, LayerParams& dst, Changes& changes);
 
     bool MergeConvolutionOrOtherAndActivation(const LayerParams& src, size_t index, QuantizationMethod method, LayerParams& dst, Changes& changes);
