+++ src/Cvt/Optimizer/Optimizer.h
@@ -46,357 +46,6 @@ namespace Synet
 
         bool OptimizeLayers(Synet::NetworkParam& network, Bytes& bin, int stage);
 
-        bool MergePermute(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
-        {
-            if (src.size() < index + 3)
-                return false;
-            const LayerParam & s0 = src[index + 0];
-            LayerParam s1 = src[index + 1];
-            LayerParam s2 = src[index + 2];
-            if (s0.type() != LayerTypePermute || s0.permute().order() != Shp(0, 3, 1, 2) && s0.permute().format() != TensorFormatNchw)
-                return false;
-            if (s1.type() != LayerTypeReshape || s1.reshape().shape().size() != 5)
-                return false;
-            if (s2.type() != LayerTypePermute || s2.permute().order() != Shp(0, 1, 3, 4, 2))
-                return false;
-            if (InsideLink(src, index + 1, 3))
-                return false;
-            const Shape & s = s1.reshape().shape();
-            s1.src() = s0.src();
-            s1.reshape().shape() = Shp(s[0], s[3], s[4], s[1], s[2]);
-            dst.push_back(s1);
-            s2.permute().format() = TensorFormatNchw;
-            s2.permute().order() = Shp(0, 3, 1, 2, 4);
-            dst.push_back(s2);
-            index += 2;
-            return true;
-        }
-
-        bool MergePooling(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
-        {
-            if (src.size() < index + 5)
-                return false;
-            if (src[index + 0].type() != LayerTypeReshape)
-                return false;
-            if (src[index + 1].type() != LayerTypePooling || src[index + 1].src()[0] != src[index + 0].name() || src[index + 1].pooling().kernel()[1] != 1)
-                return false;
-            if (src[index + 2].type() != LayerTypeReshape || src[index + 2].src()[0] != src[index + 1].name())
-                return false;
-            if (src[index + 3].type() != LayerTypeReshape || src[index + 3].src()[0] != src[index + 2].name())
-                return false;
-            if (src[index + 4].type() != LayerTypePooling || src[index + 4].src()[0] != src[index + 3].name() || src[index + 4].pooling().kernel()[1] != 1)
-                return false;
-            if (InsideLink(src, index + 1, 4))
-                return false;
-
-            LayerParam layer;
-            layer.type() = LayerTypePooling;
-            layer.name() = src[index + 4].name();
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            layer.pooling().method() = src[index + 4].pooling().method();
-            layer.pooling().kernel() = Shape({ src[index + 1].pooling().kernel()[0], src[index + 4].pooling().kernel()[0] });
-            layer.pooling().pad() = src[index + 4].pooling().pad();
-            layer.pooling().stride() = src[index + 4].pooling().stride();
-            layer.pooling().excludePad() = src[index + 4].pooling().excludePad();
-            dst.push_back(layer);
-            index += 4;
-            return true;
-        }
-
-        bool MergeSpaceToDepth(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
-        {
-            if (src.size() < index + 9)
-                return false;
-            if (src[index + 0].type() != LayerTypeStridedSlice)
-                return false;
-            if (src[index + 1].type() != LayerTypeStridedSlice)
-                return false;
-            if (src[index + 2].type() != LayerTypeStridedSlice)
-                return false;
-            if (src[index + 3].type() != LayerTypeStridedSlice)
-                return false;
-            if (src[index + 4].type() != LayerTypeStridedSlice)
-                return false;
-            if (src[index + 5].type() != LayerTypeStridedSlice)
-                return false;
-            if (src[index + 6].type() != LayerTypeStridedSlice)
-                return false;
-            if (src[index + 7].type() != LayerTypeStridedSlice)
-                return false;
-            if (src[index + 8].type() != LayerTypeConcat)
-                return false;
-            if (InsideLink(src, index + 1, 8))
-                return false;
-
-            LayerParam layer;
-            layer.type() = LayerTypeSpaceToDepth;
-            layer.name() = src[index + 8].name();
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            dst.push_back(layer);
-            index += 8;
-            return true;
-        }
-
-        bool MergeNormalize(const LayerParams & src, size_t & index, LayerParams & dst, Changes & changes)
-        {
-            if (src.size() < index + 4)
-                return false;
-            if (src[index + 0].type() != LayerTypeReduction || src[index + 0].reduction().type() != ReductionTypeL2)
-                return false;
-            if (src[index + 1].type() != LayerTypeRestrictRange || src[index + 1].restrictRange().lower() >= 0.0000001f)
-                return false;
-            if (src[index + 2].type() != LayerTypeMeta || src[index + 2].meta().type() != MetaTypeShape)
-                return false;
-            if (src[index + 3].type() != LayerTypeTile)
-                return false;
-            if (src[index + 4].type() != LayerTypeBinaryOperation || src[index + 4].binaryOperation().type() != BinaryOperationTypeDiv)
-                return false;
-
-            LayerParam layer;
-            layer.type() = LayerTypeNormalize;
-            layer.name() = src[index + 4].name();
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            layer.normalize().acrossSpatial() = true;
-            layer.normalize().channelShared() = true;
-            layer.normalize().eps() = 0;
-            dst.push_back(layer);
-            index += 4;
-            return true;
-        }
-
-        bool MergeNormalizeV2(const LayerParams& src, size_t& index, bool isNhwc, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 8)
-                return false;
-            if (src[index + 0].type() != LayerTypeReduction || src[index + 0].reduction().type() != ReductionTypeMean)
-                return false;
-            if (src[index + 1].type() != LayerTypeBinaryOperation || src[index + 1].binaryOperation().type() != BinaryOperationTypeSub)
-                return false;
-            if (src[index + 2].type() != LayerTypePower || src[index + 2].power().power() != 2.0f)
-                return false;
-            if (src[index + 3].type() != LayerTypeReduction || src[index + 3].reduction().type() != ReductionTypeMean)
-                return false;
-            if (src[index + 4].type() != LayerTypePower || src[index + 4].power().power() != 1.0f || src[index + 4].power().scale() != 1.0f)
-                return false;
-            if (src[index + 5].type() != LayerTypeUnaryOperation || src[index + 5].unaryOperation().type() != UnaryOperationTypeSqrt)
-                return false;
-            if (src[index + 6].type() != LayerTypeBinaryOperation || src[index + 6].binaryOperation().type() != BinaryOperationTypeDiv)
-                return false;
-            const WeightParam* scale = GetEltwiseWeight(index + 7, src);
-            if (scale == NULL || src[index + 7].eltwise().operation() != EltwiseOperationTypeProduct)
-                return false;
-            const WeightParam* shift = GetEltwiseWeight(index + 8, src);
-            if (shift == NULL || !IsAdd(src[index + 8]))
-                return false;
-            if (InsideLink(src, index + 1, 7))
-                return false;
-
-            LayerParam layer;
-            layer.type() = LayerTypeNormalize;
-            layer.name() = src[index + 8].name();
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            layer.normalize().eps() = src[index + 4].power().shift();
-            layer.normalize().version() = 2;
-            layer.weight().push_back(*scale);
-            layer.weight().push_back(*shift);
-            if (isNhwc && !PermutedToNchw(src, layer.src(), false, false, false))
-                layer.normalize().axis() = 1;
-            else
-                layer.normalize().axis() = -1;
-            dst.push_back(layer);
-            index += 8;
-            return true;
-        }
-
-        bool MergeNormalizeV4(const LayerParams& src, size_t& index, bool isNhwc, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 9)
-                return false;
-            if (src[index + 0].type() != LayerTypeUnaryOperation || src[index + 0].unaryOperation().type() != UnaryOperationTypeAbs)
-                return false;
-            if (src[index + 1].type() != LayerTypePower || src[index + 1].power().power() != 2.0f)
-                return false;
-            if (src[index + 2].type() != LayerTypeReduction || src[index + 2].reduction().type() != ReductionTypeSum)
-                return false;
-            if (src[index + 3].type() != LayerTypePower || src[index + 3].power().power() != 0.5f)
-                return false;
-            if (src[index + 4].type() != LayerTypeReduction || src[index + 4].reduction().type() != ReductionTypeMean)
-                return false;
-            if (src[index + 5].type() != LayerTypePower || src[index + 5].power().power() != 1.0f || src[index + 5].power().scale() != 1.0f)
-                return false;
-            if (src[index + 6].type() != LayerTypeBinaryOperation || src[index + 6].binaryOperation().type() != BinaryOperationTypeDiv)
-                return false;
-            if (src[index + 7].type() != LayerTypeEltwise || src[index + 7].eltwise().operation() != EltwiseOperationTypeProduct)
-                return false;
-            if (src[index + 8].type() != LayerTypeScale || !src[index + 8].scale().biasTerm())
-                return false;
-            if (!IsAdd(src[index + 9]))
-                return false;
-            if (InsideLink(src, index + 1, 8))
-                return false;
-            LayerParam layer;
-            layer.type() = LayerTypeNormalize;
-            layer.name() = src[index + 9].name();
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            layer.normalize().eps() = src[index + 5].power().shift();
-            layer.normalize().version() = 4;
-            layer.weight() = src[index + 8].weight();
-            if (isNhwc && !PermutedToNchw(src, layer.src(), false, false, false))
-                layer.normalize().axis() = -1;
-            else
-                layer.normalize().axis() = 1;
-            dst.push_back(layer);
-            index += 9;
-            return true;
-        }
-
-        bool MergeNormalizeV5(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 4)
-                return false;
-            if (src[index + 0].type() != LayerTypeReshape || src[index + 0].reshape().shape().size() != 3)
-                return false;
-            if (src[index + 1].type() != LayerTypeNormalize || src[index + 1].normalize().version() != 3 || src[index + 1].src()[0] != src[index + 0].dst()[0])
-                return false;
-            if (src[index + 2].type() != LayerTypeMeta || src[index + 2].meta().type() != MetaTypeShape || src[index + 2].src()[0] != src[index + 0].src()[0])
-                return false;
-            if (src[index + 3].type() != LayerTypeReshape || src[index + 3].src()[0] != src[index + 1].dst()[0] || src[index + 3].src()[1] != src[index + 2].dst()[0])
-                return false;
-
-            LayerParam layer;
-            layer.type() = LayerTypeNormalize;
-            layer.name() = src[index + 3].name();
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            layer.normalize() = src[index + 1].normalize();
-            layer.normalize().version() = 5;
-            layer.weight() = src[index + 1].weight();
-            dst.push_back(layer);
-            index += 3;
-            return true;
-        }
-
-        bool MergeScale(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 2)
-                return false;
-            const WeightParam* scale = GetEltwiseWeight(index + 0, src);
-            if (scale == NULL || src[index + 0].eltwise().operation() != EltwiseOperationTypeProduct)
-                return false;
-            const WeightParam* shift = GetEltwiseWeight(index + 1, src);
-            if (shift == NULL || !IsAdd(src[index + 1]))
-                return false;
-            if(src[index + 1].src()[0] != src[index + 0].dst()[0])
-                return false;
-            if (scale->dim() != shift->dim())
-                return false;
-
-            LayerParam layer;
-            layer.type() = LayerTypeScale;
-            layer.name() = src[index + 1].name();
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            layer.scale().biasTerm() = true;
-            layer.weight().push_back(*scale);
-            layer.weight().push_back(*shift);
-            dst.push_back(layer);
-            index += 1;
-            return true;
-        }
-
-        bool MergeTiledScale2D(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 4)
-                return false;
-            if (src[index + 0].type() != LayerTypeTile || src[index + 0].src().size() != 2)
-                return false;
-            if (src[index + 1].type() != LayerTypeTile || src[index + 1].src().size() != 2)
-                return false;
-            if (!IsMul(src[index + 2]) || src[index + 2].src().size() != 2)
-                return false;
-            if (!IsMul(src[index + 3]) || src[index + 3].src().size() != 2)
-                return false;
-            if (src[index + 0].src()[1] != src[index + 1].src()[1] || 
-                src[index + 2].src()[1] != src[index + 1].dst()[0] ||
-                src[index + 3].src()[1] != src[index + 0].dst()[0])
-                return false;
-
-            LayerParam layer;
-            layer.type() = LayerTypeTiledScale2D;
-            layer.name() = src[index + 3].name();
-            layer.src().push_back(src[index + 2].src()[0]);
-            layer.src().push_back(src[index + 1].src()[0]);
-            layer.src().push_back(src[index + 0].src()[0]);
-            layer.dst().push_back(layer.name());
-            dst.push_back(layer);
-            index += 3;
-            return true;
-        }
-
-        bool MergeUnpack4(const LayerParams& src, size_t& index, bool isNhwc, LayerParams& dst, Changes& changes)
-        {
-            if (src.size() < index + 19)
-                return false;
-            if (src[index + 0].type() != LayerTypeMeta || src[index + 0].meta().type() != MetaTypeShape)
-                return false;
-            if (!IsMetaConst64i(src[index + 1]))
-                return false;
-            if (src[index + 2].type() != LayerTypeMeta || src[index + 2].meta().type() != MetaTypeGather)
-                return false;
-            if (!IsMetaConst64i(src[index + 3]))
-                return false;
-            if (!IsMetaConst64i(src[index + 4]))
-                return false;
-            if (src[index + 5].type() != LayerTypeMeta || src[index + 5].meta().type() != MetaTypeAdd)
-                return false;
-            if (!IsMetaConst64i(src[index + 6], Lng(4)))
-                return false;
-            if (src[index + 7].type() != LayerTypeMeta || src[index + 7].meta().type() != MetaTypeDiv)
-                return false;
-            if (!IsMetaConst64i(src[index + 8], Lng(1)))
-                return false;
-            if (src[index + 9].type() != LayerTypeMeta || src[index + 9].meta().type() != MetaTypeMul)
-                return false;
-            if (src[index + 10].type() != LayerTypeStridedSlice || src[index + 10].src()[0] != src[index + 0].src()[0])
-                return false;
-            if (!IsMetaConst64i(src[index + 11], Lng(2)))
-                return false;
-            if (src[index + 12].type() != LayerTypeMeta || src[index + 12].meta().type() != MetaTypeMul)
-                return false;
-            if (src[index + 13].type() != LayerTypeStridedSlice || src[index + 13].src()[0] != src[index + 0].src()[0])
-                return false;
-            if (!IsMetaConst64i(src[index + 14], Lng(3)))
-                return false;
-            if (src[index + 15].type() != LayerTypeMeta || src[index + 15].meta().type() != MetaTypeMul)
-                return false;
-            if (src[index + 16].type() != LayerTypeStridedSlice || src[index + 16].src()[0] != src[index + 0].src()[0])
-                return false;
-            if (!IsMetaConst64i(src[index + 17], Lng(4)))
-                return false;
-            if (src[index + 18].type() != LayerTypeMeta || src[index + 18].meta().type() != MetaTypeMul)
-                return false;
-            if (src[index + 19].type() != LayerTypeStridedSlice || src[index + 19].src()[0] != src[index + 0].src()[0])
-                return false;
-
-            LayerParam layer;
-            layer.type() = LayerTypeUnpack;
-            layer.name() = src[index + 0].name();
-            layer.src() = src[index + 0].src();
-            layer.unpack().axis() = isNhwc ? 3 : 1;
-            layer.dst().push_back(src[index + 10].dst()[0]);
-            layer.dst().push_back(src[index + 13].dst()[0]);
-            layer.dst().push_back(src[index + 16].dst()[0]);
-            layer.dst().push_back(src[index + 19].dst()[0]);
-            dst.push_back(layer);
-            index += 19;
-            return true;
-        }
-
         bool MergeRnnGruBd(const LayerParams& src, size_t& index, LayerParams& dst, Changes& changes)       
         {
             const size_t RNN_GRU_BD_SIZE = 19;
@@ -676,20 +325,6 @@ namespace Synet
         }
 
         //-------------------------------------------------------------------------------------------------
-
-        const WeightParam* GetEltwiseWeight(size_t index, const LayerParams& layers) const
-        {
-            if (index < layers.size() && (layers[index].type() == LayerTypeEltwise && layers[index].src().size() == 2) || layers[index].type() == LayerTypeAdd)
-            {
-                const LayerParam* src0 = GetLayer(layers, layers[index].src()[0]);
-                if (src0 && src0->type() == LayerTypeConst)
-                    return src0->weight().data() + 0;
-                const LayerParam* src1 = GetLayer(layers, layers[index].src()[1]);
-                if (src1 && src1->type() == LayerTypeConst)
-                    return src1->weight().data() + 0;
-            }
-            return NULL;
-        }
 
         bool Rename(const Change & change, LayerParams & layers)
         {
